added extra functions to cryptofaucet and referred to them in the scripts
