+++ bin/bitgo_viewwallets.js
@@ -1,19 +1,12 @@
-var bitgo = require("../lib/bitgoclient");
+var cryptofaucet = require("../app/cryptofaucet");
 
-
-var wallets = bitgo.coin('tbtc').wallets();
-wallets.list({}, function callback(err, data) {
-// handle error, do something with wallets
-    for (var id in data.wallets) {
-        //console.dir("data.wallets["+id+"] = "+JSON.stringify(data.wallets[id]));
-        var wallet = data.wallets[id]._wallet;
-        var shortenedObj = {
-            "id":wallet.id,
-            "label":wallet.label,
-            //"latest_address":wallet.receiveAddress.address,
-            "balance":wallet.balance
-        };
-        console.log(shortenedObj);
+var cryptoSymbol = "tbtc";
+cryptofaucet.getWallets(cryptoSymbol,(err,walletsList) => {
+    if (err) {
+        console.log("err = "+err);
+        onComplete(err);
     }
+    console.log(walletsList);
 });
+
